+++ asm/nonmatchings/code_8800/create_cameras_a.s
