+++ asm/nonmatchings/code_8800/initialize_next_camera.s
