+++ asm/nonmatchings/code_B0E0/do_camera_type_3.s
