+++ src/code_B0E0.c
@@ -1,0 +1,18 @@
+#include "common.h"
+
+INCLUDE_ASM(s32, "code_B0E0", do_camera_type_unused);
+
+INCLUDE_ASM(s32, "code_B0E0", do_camera_type_5);
+
+INCLUDE_ASM(s32, "code_B0E0", func_80030210);
+
+INCLUDE_ASM(s32, "code_B0E0", func_8003034C);
+
+void func_80030450(void) {
+}
+
+INCLUDE_ASM(s32, "code_B0E0", test_ray_zone_aabb);
+
+INCLUDE_ASM(s32, "code_B0E0", func_800304FC);
+
+INCLUDE_ASM(s32, "code_B0E0", do_camera_type_3);
