+++ src/code_ad90_len_2880.c
@@ -1,20 +1,3 @@
 #include "common.h"
 
 INCLUDE_ASM(s32, "code_ad90_len_2880", do_camera_type_6);
-
-INCLUDE_ASM(s32, "code_ad90_len_2880", do_camera_type_unused);
-
-INCLUDE_ASM(s32, "code_ad90_len_2880", do_camera_type_5);
-
-INCLUDE_ASM(s32, "code_ad90_len_2880", func_80030210);
-
-INCLUDE_ASM(s32, "code_ad90_len_2880", func_8003034C);
-
-void func_80030450(void) {
-}
-
-INCLUDE_ASM(s32, "code_ad90_len_2880", test_ray_zone_aabb);
-
-INCLUDE_ASM(s32, "code_ad90_len_2880", func_800304FC);
-
-INCLUDE_ASM(s32, "code_ad90_len_2880", do_camera_type_3);
