+++ tools/splat.yaml
@@ -44,8 +44,10 @@ segments:
     - [0x7e40, c, os/code_7e40_len_3f0]
     - [0x8230, c, os/nusys/nuSiMgr]
     - [0x8560, c]
+    - [0x8800, c]
     - [0x9d10, c, code_9d10_len_1080]
     - [0xad90, c, code_ad90_len_2880]
+    - [0xB0E0, c]
     - [0xd610, c, code_d610_len_1330]
     - [0xe940, c, code_e940_len_290]
     - [0xebd0, c, code_ebd0_len_6a0]
@@ -92,7 +94,7 @@ segments:
     - [0x3b2f0, c, os/nusys/nuContRmbCheck]
     - [0x3b390, c, os/nusys/nuContRmbModeSet]
     - [0x3b4a0, c, os/nusys/code_3B4A0]
-    - [0x3B5F0, c, os/nusys/nuGfxThread]
+    - [0x3B5F0, c, os/nusys/nuGfxThread] # End of nusys
     - [0x3b710, c, os/code_3b710_len_40]
     - [0x3b750, c, os/code_3b750_len_20]
     - [0x3b770, c, os/code_3b770_len_80]
